requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Mail>=0.10.0
email-validator>=2.1.0
WTForms>=3.1.0
Werkzeug>=3.0.0
reportlab>=4.0.0
pytz>=2024.1
python-dotenv>=1.0.0
itsdangerous>=2.1.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.0
